// packages/dirvi-lib/src/tree-surfer/fold-node/fold-node.types.d.ts
export type FoldNode<Id extends SerializableKey> = {
    id: Id;
    children: FoldNode<Id>[];
    foldedChildren: FoldNode<Id>[];
};
// ...
export type FoldNodeApi<Id extends SerializableKey> = TreeNodeApi<Id, FoldNode<Id>>;